refactor(client): restructure internal package
