+++ client/internal/eventbus/config.go
