+++ client/internal/eventlog/config.go
