+++ client/internal/eventlog/name_service.go
