+++ client/internal/store/alloc.go
