+++ client/internal/store/allocator.go
